# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API маркетплейса Клевер (микросервисная архитектура).\n- Auth Service       (HTTP :8001, gRPC :9001) — авторизация, профиль, пользователи\n- Support Service    (HTTP :8002)             — техподдержка\n- Catalog Service    (HTTP :8004, gRPC :9004) — объявления, категории, избранное, просмотры\n- Commerce Service   (HTTP :8006)             — корзина, чаты, заказы\nВсе запросы идут через API Gateway (nginx :8080).",
        "title": "Clover API",
        "contact": {},
        "version": "1.0"
    },
    "host": "clover-go.ru",
    "basePath": "/api/v1",
    "paths": {
...
        "/support/tickets": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Возвращает список обращений текущего пользователя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "support"
                ],
                "summary": "Мои обращения",
                "responses": {
                    "200": {
                        "description": "список обращений получен",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.TicketResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "unauthorized: Пользователь не авторизован",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal error: Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Создаёт новое обращение в техподдержку",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "support"
                ],
                "summary": "Создать обращение",
                "parameters": [
                    {
                        "description": "Данные обращения",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateTicketRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "обращение успешно создано",
                        "schema": {
                            "$ref": "#/definitions/dto.TicketResponse"
                        }
                    },
                    "400": {
                        "description": "bad request: Ошибка валидации или неверный JSON",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "unauthorized: Пользователь не авторизован",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal error: Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/support/tickets/{id}": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Возвращает детали одного обращения по ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "support"
                ],
                "summary": "Получить обращение",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID обращения",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "обращение получено",
                        "schema": {
                            "$ref": "#/definitions/dto.TicketResponse"
                        }
                    },
                    "400": {
                        "description": "bad request: Некорректный ID",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "unauthorized: Пользователь не авторизован",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "forbidden: Нет доступа к чужому обращению",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "not found: Обращение не найдено",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal error: Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Обновляет данные обращения (только в статусе open)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "support"
                ],
                "summary": "Обновить обращение",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID обращения",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые данные обращения",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateTicketRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "обращение успешно обновлено",
                        "schema": {
                            "$ref": "#/definitions/dto.TicketResponse"
                        }
                    },
                    "400": {
                        "description": "bad request: Ошибка валидации, неверный статус или JSON",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "unauthorized: Пользователь не авторизован",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "forbidden: Нет доступа к чужому обращению",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "not found: Обращение не найдено",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal error: Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/support/tickets/{id}/rate": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Позволяет автору обращения выставить оценку от 1 до 5 для закрытого тикета",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "support"
                ],
                "summary": "Оценить обращение",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID обращения",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Оценка",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RateTicketRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "оценка выставлена",
                        "schema": {
                            "$ref": "#/definitions/dto.TicketResponse"
                        }
                    },
                    "400": {
                        "description": "bad request: Некорректная оценка, тикет не закрыт или уже оценён",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "unauthorized: Пользователь не авторизован",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "forbidden: Пользователь не является автором",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "not found: Обращение не найдено",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal error: Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateTicketRequest": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "dto.RateTicketRequest": {
            "type": "object",
            "properties": {
                "rating": {
                    "type": "integer"
                }
            }
        },
...
        "dto.TicketResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "rating": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "dto.UpdateTicketRequest": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "CookieAuth": {
            "description": "JWT session token",
            "type": "apiKey",
            "name": "token",
            "in": "cookie"
        },
        "CsrfCookieAuth": {
            "description": "CSRF token stored in cookies",
            "type": "apiKey",
            "name": "csrf_token",
            "in": "cookie"
        },
        "CsrfHeaderAuth": {
            "description": "CSRF token required in header",
            "type": "apiKey",
            "name": "X-CSRF-Token",
            "in": "header"
        }
    }
}